rank bug fixed
