+++ src/api.js
@@ -49,7 +49,6 @@ class apiService {
             })
 
             const result = await response.json()
-            console.log(result)
 
             return result
         } catch (error) {
